Add powertest output trimming scripts
This commit adds scripts for powertest output trimming.  It receives
files containing output of each variant of tpc-h and reconstruct them in
table like format.  It is not only easier for human eye, but also can be
used as a input to plot_stdin.py of lazybox for chart generation.  For
example, you may use command like below:
```
$ ./outputs.sh nocma.sf1 cma.buddy.sf1 | \
	./outputs_table.py | \
	~/lazybox/gnuplot/plot_stdin.py clustered_box
```

Signed-off-by: SeongJae Park <sj38.park@gmail.com>